Fixed some bugs, levels are now all winnable (hopefully)
I made a new function called makeRatioPossible that sets the current
string so that the goal frequency is matchable.

I also changed the function drawRectangles so that the color of the
goal string is displayed.

I edited the if statement where the hasWon() is checked so that the
array doesn’t go out of bounds.

Still need to get rid of all the outputs