Added a play button to play the 2 strings simultaneously
There is a new fn called setSpecificValues to take a frequency and
generate correct values for tension, weight, and length. I need to add
a case so that if the specified freq is not possible, it handles the
error instead of just stalling out.

I also added more discrimination on the weight so you can get closer to
the correct frequency. I cut the error margin to 0.005