Changing current string, hiding frequency values
Can double-click on a string to make it the current string.
Current string is now displayed above the sliders.
When string is switched, sliders are reset to the current value.
Need to add function way to disable this on levels that are not free
play levels.

Freq value is not displayed until winAttempts > 5 or level>5.
winAttempts is incremented when both strings are playing simultaneously